backend/requirements.txt
Django>=3.2
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
